Extend `walk-directory` function in `pathnames` package
Replace the `condition` option with two more specific conditions:
`file-condition` and `directory-condition` that make life much easier for users
of this function.

Prior to this, if they wanted to avoid going into a directory while traversing
recursively, they would have to write `condition` as a more complex expression.
Plus, it was simply harder to intuit how `condition` worked for both files and
directories.