Fix compilation error in `binary.lisp`
A macro was calling `assert-all` as part of its implementation (and not just as part
of the expansion code), which resulted in an error because even though the
function was defined earlier in the file, it wasn't being available during
compilation time, a condition necessary for macros trying to use user functions
like that one.

Wrapping the function definition in a `eval-when` clause fixes the problem.