Support sorting of frames analysis by tag version and frame ID
Even though we don't have actual support for v2.4, we want to see if we can
glean some information from mp3s in that version using the code for v2.3.

According to
https://mutagen-specs.readthedocs.io/en/latest/id3/id3v2.4.0-structure.html
(which to be honest, I haven't read in detail yet) there was an effort to keep
backward compatibility, so this change is hopefully not totally useless.

Running

(frame-types-in-directory "music" :sort-by 'id)
(frame-types-in-directory "music" :sort-by 'version)

actually returns results and doesn't report any failures to parse anything, so
there is that.