Better indent `define-binary-type` macro
Although maybe it wasn't a big deal, I found that the indentation of
the (:reader ...) and (:writer ...) clauses in `define-binary-type` strongly
suggested that we were making function calls, as opposed to defining some
structure, so I did some digging and found a way to customize indentation using
the package `trivial-indent`, which works by giving hints to SLIME about the
desired indentation.

It's not as flexible as I would like it, but it gets the job done here.