Added `with-labels` as a very thin wrapper for `labels`
While it's a trivial wrapper, I think it improves readability by allowing one to
focus first on the main expression before the auxiliary definitions.

One limitation is that unlike `labels`, this form only accepts a single
expression, but I find that this is by far the most common form that I tend to
use, and there are easy workarounds for cases where I want multiple expressions
executed.