Bug Fix: playlist page is not rendering properly
The culprit was a silent error in the rendering of a FOO macro.
Unfortunately, the FOO library doesn't do much in terms of tag
validation, so when a macro is used but there is no definition
available, the tag is rendered as if it were just a regular HTML tag,
generating nonsense HTML.

The macro in question (`:mp3-browser-page`) wasn't available because
the definitions were being loaded in incorrect order (this was a
regression introduced in a previous commit.)

The fix is simply to separate the UI templates and have other files
depend on them in the ASDF definition file.